Change to webpack emit instead of direct write.
Change to webpack 'emit' instead of writing directly to the filesystem. Doing it this way has the advantage that the emitted files get recognised by other plugins (e.g. [offline-plugin](https://github.com/NekR/offline-plugin)) and are listed in the webpack log.

This introduces a (breaking) change to the config options. Files now get writen to the output path set in the webpack options. The targetDir option now sets a directory in the output path and has to end with a '/' (at least for now).

Since the options related change is breaking I bumped the version to 2.0.